Fix backwards compatibility issues.
The extension was compiling against the VS2015 version of the 2012
shell, which caused the extension to fail install on VS2012 and VS2013
if VS2015 was not also installed. Fixed by specifying an older version
number to compile against. Also added the VS2012 isolated shell
installer as a redistributable so you can compile the extension with
VS2015 without having to have VS2012 installed as well.