Fix not being able to reflow Python docstrings
Blocks whose block start and block end patterns were the same and which did not have a line start pattern could not be reflowed when followed by a blank line. Any trailing blank line would be interpreted as a line start, causing the block end to be interpreted as a block start. Fixed by not attempting to infer a block start if there is no line start pattern.